Fix baseUrl handling
The baseUrl did not handle trailing slashes correctly.
The V2 Client expected the baseUrl to not have a trailing slash.
The V1 Client expected the baseUrl to have a trailing slash.

With the fix the trailing slash is always removed and added during the
calculation of the subscription path.

Fixes #99